feat: `impl Schema for PathBuf`
NOTE: This is a very quick and "mindless" implementation. I think codegens should be able to serialize it however they want to, not just as a string.

For that reason, we probably need to provide "type id" of the original type, since we can't just create "special case primitives" for all types.